Isolate WebSocket subscription policy
